requirements.txt
fastapi
uvicorn
torch
pillow
pdf2image
python-multipart
